Add "Copy web URL" command to the Copy context menu. (#16415)
This command copies a web deeplink url onto clipboard, which is useful
when SiYuan is running in browser and the link will use the format
of "https://app.com/?id={id}". This link can be then easily shared
to other users who have access to the same instance.

The item is only visible in browsers.